rename component, add global classes
