+++ packages/react/ds/src/combo-box/types.ts
@@ -7,7 +7,7 @@ export type DropdownItemType = {
   noSearch?: boolean;
 };
 
-export type MultiSelectSearchProps = {
+export type ComboboxProps = {
   // use any in order to use ServerActions for NextJs applications
   action: any; // eslint-disable-line @typescript-eslint/no-explicit-any
   dropdownItems: DropdownItemType[];
