--- packages/react/ds/src/multi-select-search/multi-select-search.stories.tsx
@@ -1,726 +1,0 @@
-import { Meta, StoryObj } from '@storybook/react';
-import { MultiSelectSearch } from './multi-select-search.js';
-
-const meta = {
-  title: 'Navigation/MultiSelectSearch',
-  parameters: {
-    description: {
-      component: 'ToDO',
-    },
-  },
-  component: MultiSelectSearch,
-} satisfies Meta<typeof MultiSelectSearch>;
-
-export default meta;
-type Story = StoryObj<typeof meta>;
-
-export const Default: Story = {
-  args: {
-    action: '#',
-    dropdownItems: [
-      {
-        label: 'Organisations',
-        options: [
-          {
-            label: 'An Bord Pleanála',
-            value: '15431907-an-bord-pleanala',
-          },
-          {
-            label: 'An Garda Síochána',
-            value: 'an-garda-siochana',
-          },
-          {
-            label: 'Bord Bia (Irish Food Board)',
-            value: '14061907-bord-bia-irish-food-board',
-          },
-          {
-            label: 'Capital Works Management Framework',
-            value: '7c2ea-capital-works-management-framework',
-          },
-          {
-            label: 'Carlow County Council',
-            value: 'carlow-county-council',
-          },
-          {
-            label: 'Cavan County Council',
-            value: 'cavan-county-council',
-          },
-          {
-            label: 'Central Bank of Ireland',
-            value: '10032007-central-bank-of-ireland',
-          },
-          {
-            label: 'Central Statistics Office',
-            value: 'central-statistics-office',
-          },
-          {
-            label: 'Chief State Solicitor’s Office',
-            value: 'chief-state-solicitors-office',
-          },
-          {
-            label: 'Circulars',
-            value: '28448-circulars',
-          },
-          {
-            label: 'Civil Justice Review',
-            value: 'dff8b-civil-justice-review',
-          },
-          {
-            label: 'Civil Service Employee Assistance Service',
-            value: '72752-civil-service-employee-assistance-service',
-          },
-          {
-            label: 'Clare County Council',
-            value: 'clare-county-council',
-          },
-          {
-            label: 'Commission on Taxation and Welfare',
-            value: 'c526e-commission-on-taxation-and-welfare',
-          },
-          {
-            label: 'Commission on the Defence Forces',
-            value: '20d04-commission-of-the-defence-forces',
-          },
-          {
-            label: 'Competition and Consumer Protection Commission (CCPC)',
-            value:
-              '15341907-competition-and-consumer-protection-commission-ccpc',
-          },
-          {
-            label: 'Cork City Council',
-            value: 'cork-city-council',
-          },
-          {
-            label: 'Cork County Council',
-            value: 'cork-county-council',
-          },
-          {
-            label: 'Coroner',
-            value: 'da922-coroner',
-          },
-          {
-            label: 'Courts Service',
-            value: 'the-courts-service',
-          },
-          {
-            label: 'Criminal Injuries Compensation Scheme',
-            value: '90ab1-criminal-injuries-compensation-scheme',
-          },
-          {
-            label: 'Criminal Law Codification Advisory Committee',
-            value: '8c220-criminal-law-codification-advisory-committee',
-          },
-          {
-            label: 'Data Protection Commission',
-            value: '14521907-data-protection-commission',
-          },
-          {
-            label: 'Department of Agriculture, Food and the Marine',
-            value: 'department-of-agriculture-food-and-the-marine',
-          },
-          {
-            label:
-              'Department of Children, Equality, Disability, Integration and Youth',
-            value:
-              'department-of-children-equality-disability-integration-and-youth',
-          },
-          {
-            label: 'Department of Defence',
-            value: 'department-of-defence',
-          },
-          {
-            label: 'Department of Education',
-            value: 'department-of-education',
-          },
-          {
-            label: 'Department of Enterprise, Trade and Employment',
-            value: 'department-of-enterprise-trade-and-employment',
-          },
-          {
-            label: 'Department of Finance',
-            value: 'department-of-finance',
-          },
-          {
-            label: 'Department of Foreign Affairs',
-            value: 'department-of-foreign-affairs',
-          },
-          {
-            label:
-              'Department of Further and Higher Education, Research, Innovation and Science',
-            value: 'department-of-higher-education-innovation-and-science',
-          },
-          {
-            label: 'Department of Health',
-            value: 'department-of-health',
-          },
-          {
-            label: 'Department of Housing, Local Government and Heritage',
-            value: 'department-of-housing-local-government-and-heritage',
-          },
-          {
-            label: 'Department of Justice',
-            value: 'department-of-justice',
-          },
-          {
-            label: 'Department of Public Expenditure, NDP Delivery and Reform',
-            value: 'department-of-public-expenditure-and-reform',
-          },
-          {
-            label: 'Department of Rural and Community Development',
-            value: 'department-of-rural-and-community-development',
-          },
-          {
-            label: 'Department of Social Protection',
-            value: 'department-of-social-protection',
-          },
-          {
-            label: 'Department of the Environment, Climate and Communications',
-            value: 'department-of-the-environment-climate-and-communications',
-          },
-          {
-            label: 'Department of the Taoiseach',
-            value: 'department-of-the-taoiseach',
-          },
-          {
-            label:
-              'Department of Tourism, Culture, Arts, Gaeltacht, Sport and Media',
-            value:
-              'department-of-tourism-culture-arts-gaeltacht-sport-and-media',
-          },
-          {
-            label: 'Department of Transport',
-            value: 'department-of-transport',
-          },
-          {
-            label: 'Digital Hub Development Agency',
-            value: '4e9b9-digital-hub-development-agency',
-          },
-          {
-            label: 'Digital postbox',
-            value: 'fc836-digital-postbox',
-          },
-          {
-            label: 'Donegal County Council',
-            value: 'donegal-county-council',
-          },
-          {
-            label: 'Dublin City Council',
-            value: 'dublin-city-council',
-          },
-          {
-            label: 'Dún Laoghaire–Rathdown County Council',
-            value: 'dun-laoghaire-rathdown-county-council',
-          },
-          {
-            label: 'Education Scoping Inquiry',
-            value: '6e681-education-scoping-inquiry',
-          },
-          {
-            label: 'Educational Research Centre (ERC)',
-            value: '09082007-educational-research-centre-erc',
-          },
-          {
-            label: 'EGovernment Strategy',
-            value: '021aa-egovernment-strategy',
-          },
-          {
-            label: 'Embassy of Ireland, Argentina',
-            value: 'e6fbc-embassy-of-ireland-argentina',
-          },
-          {
-            label: 'Enterprise Ireland',
-            value: '15331907-enterprise-ireland',
-          },
-          {
-            label: 'European Structural and Investment Funds',
-            value: '44633-european-structural-and-investment-funds',
-          },
-          {
-            label: 'External Oversight Body of the Defence Forces',
-            value: '0b494-external-oversight-body-of-the-defence-forces',
-          },
-          {
-            label: 'Financial Services and Pensions Ombudsman (FSPO)',
-            value: '16481907-financial-services-and-pensions-ombudsman-fspo',
-          },
-          {
-            label: 'Fingal County Council',
-            value: 'fingal-county-council',
-          },
-          {
-            label: 'Forensic Science Ireland',
-            value: '07c21-forensic-science-ireland',
-          },
-          {
-            label: 'Freedom of Information',
-            value: '57b45-freedom-of-information',
-          },
-          {
-            label: 'Galway City Council',
-            value: 'galway-city-council',
-          },
-          {
-            label: 'Galway County Council',
-            value: 'galway-county-council',
-          },
-          {
-            label: 'Gambling Regulatory Authority of Ireland',
-            value: 'dbd5e-gambling-regulatory-authority-of-ireland',
-          },
-          {
-            label: 'Garda Inspectorate',
-            value: '04c91-garda-inspectorate',
-          },
-          {
-            label: 'Gender Equality',
-            value: 'f1304-gender-equality',
-          },
-          {
-            label: 'Gov.ie news',
-            value: 'govie-news',
-          },
-          {
-            label: 'Health Research Board (HRB)',
-            value: '16151907-health-research-board-hrb',
-          },
-          {
-            label: 'Health Service Executive',
-            value: '12361907-health-service-executive',
-          },
-          {
-            label: 'Healthy Ireland',
-            value: 'healthy-ireland-',
-          },
-          {
-            label: 'Higher Education Authority (HEA)',
-            value: '17191907-higher-education-authority',
-          },
-          {
-            label: 'IDA Ireland',
-            value: '15301907-ida-ireland',
-          },
-          {
-            label: 'Independent Review Group - Defence Forces',
-            value: '68e8b-independent-review-group-defence-forces',
-          },
-          {
-            label: 'Insolvency Service of Ireland',
-            value: '32f6d-insolvency-service-of-ireland',
-          },
-          {
-            label: 'International Protection Accommodation Services (IPAS)',
-            value: 'd634d-international-protection-accommodation-services-ipas',
-          },
-          {
-            label: 'Irish Fiscal Advisory Council',
-            value: '15011607-irish-fiscal-advisory-council',
-          },
-          {
-            label: 'Irish Government Economic and Evaluation Service (IGEES)',
-            value:
-              '99896-irish-government-economic-and-evaluation-service-igees',
-          },
-          {
-            label: 'Irish Prison Service',
-            value: 'irish-prison-service',
-          },
-          {
-            label: 'Irish Treaty Series',
-            value: '1bc60-irish-treaty-series',
-          },
-          {
-            label: 'Kerry County Council',
-            value: 'kerry-county-council',
-          },
-          {
-            label: 'Kildare County Council',
-            value: 'kildare-county-council',
-          },
-          {
-            label: 'Kilkenny County Council',
-            value: 'kilkenny-county-council',
-          },
-          {
-            label: 'Laois County Council',
-            value: 'laois-county-council',
-          },
-          {
-            label: 'Legal Practitioners Disciplinary Tribunal',
-            value: '7a62c-legal-practitioners-disciplinary-tribunal',
-          },
-          {
-            label: 'Leitrim County Council',
-            value: 'leitrim-county-council',
-          },
-          {
-            label: 'Limerick City and County Council',
-            value: 'limerick-city-county-council',
-          },
-          {
-            label: 'Local Government Management Agency (LGMA)',
-            value: '15371907-local-government-management-agency-lgma',
-          },
-          {
-            label: 'Longford County Council',
-            value: 'longford-county-council',
-          },
-          {
-            label: 'Louth County Council',
-            value: 'louth-county-council',
-          },
-          {
-            label: 'Mayo County Council',
-            value: 'mayo-county-council',
-          },
-          {
-            label: 'Meath County Council',
-            value: 'meath-county-council',
-          },
-          {
-            label: 'Monaghan County Council',
-            value: 'monaghan-county-council',
-          },
-          {
-            label: 'National Asset Management Agency (NAMA)',
-            value: '16441907-nama-national-asset-management-agency',
-          },
-          {
-            label: 'National Shared Services Office',
-            value: 'bf6cf8-the-national-shared-services-office',
-          },
-          {
-            label: 'National Treasury Management Agency (NTMA)',
-            value: '16411907-national-treasury-management-agency-ntma',
-          },
-          {
-            label: 'Offaly County Council',
-            value: 'offaly-county-council',
-          },
-          {
-            label: 'Office of Emergency Planning',
-            value: 'ec39cf-office-of-emergency-planning',
-          },
-          {
-            label: 'Office of Government Procurement',
-            value: 'office-of-government-procurement',
-          },
-          {
-            label: 'Office of Public Works',
-            value: 'office-of-public-works',
-          },
-          {
-            label: 'Office of the Attorney General',
-            value: 'office-of-the-attorney-general',
-          },
-          {
-            label: 'Office of the Civil Service Chief Medical Officer',
-            value: 'ee1ba-office-of-the-chief-medical-officer',
-          },
-          {
-            label: 'Office of the Comptroller and Auditor General',
-            value: 'office-of-the-comptroller-and-auditor-general',
-          },
-          {
-            label: 'Office of the Director of Authorised Intervention, Tuam',
-            value:
-              'a8167-office-of-the-director-of-authorised-intervention-tuam',
-          },
-          {
-            label: 'Office of the Planning Regulator',
-            value: '488240-office-of-the-planning-regulator',
-          },
-          {
-            label: 'Office of the Revenue Commissioners',
-            value: 'office-of-the-revenue-commissioners',
-          },
-          {
-            label: 'Office of the State Pathologist',
-            value: '3cc86-office-of-the-state-pathologist',
-          },
-          {
-            label: 'Parliamentary Budget Office',
-            value: '82e88-parliamentary-budget-office',
-          },
-          {
-            label: 'Parole Board',
-            value: 'parole-board',
-          },
-          {
-            label: 'Pobal',
-            value: '15361907-pobal',
-          },
-          {
-            label: " President's Establishment ",
-            value: 'presidents-establishment',
-          },
-          {
-            label: 'Probation Service',
-            value: '25cc7-probation-service',
-          },
-          {
-            label: 'Public Appointments Service',
-            value: 'public-appointments-service',
-          },
-          {
-            label: 'Public Consultations',
-            value: '2f69a-public-consultations',
-          },
-          {
-            label: 'Public Private Partnership',
-            value: '55a37-public-private-partnership',
-          },
-          {
-            label: 'Public Service ICT Strategy',
-            value: '6e74c-public-service-ict-strategy',
-          },
-          {
-            label: 'Public Service Pay Commission',
-            value: 'fa328-public-service-pay-commission',
-          },
-          {
-            label: 'Public Service Reform Plan',
-            value: 'afaea-public-service-reform-plan',
-          },
-          {
-            label: 'Publicjobs',
-            value: 'aea02-publicjobs',
-          },
-          {
-            label: 'Roscommon County Council',
-            value: 'roscommon-county-council',
-          },
-          {
-            label: 'Safety on the Water',
-            value: 'f6188-safety-on-the-water',
-          },
-          {
-            label: 'Single Public Service Pension Scheme',
-            value: 'e0a4f-single-public-service-pension-scheme',
-          },
-          {
-            label: 'Sligo County Council',
-            value: 'sligo-county-council',
-          },
-          {
-            label: 'Social Welfare Appeals Office',
-            value: '4d5227-social-welfare-appeals-office',
-          },
-          {
-            label: 'South Dublin County Council',
-            value: 'south-dublin-county-council',
-          },
-          {
-            label: 'Special EU Programmes Body',
-            value: '44b19-special-eu-programmes-body',
-          },
-          {
-            label: 'State Examinations Commission (SEC)',
-            value: '16521907-state-examinations-commission-sec',
-          },
-          {
-            label: 'Strategic Banking Corporation of Ireland',
-            value: '16381907-strategic-banking-corporation-of-ireland',
-          },
-          {
-            label: 'Study on Familicide and Domestic Homicide Reviews',
-            value: 'eecd0-study-on-familicide-and-domestic-homicide-reviews',
-          },
-          {
-            label: 'Tailte Éireann',
-            value: '69ad4-tailte-eireann',
-          },
-          {
-            label: 'The Charities Regulator',
-            value: '14561907-charities-regulatory-authority',
-          },
-          {
-            label: 'The Funds Administration Unit',
-            value: '9597c-the-funds-administration-unit',
-          },
-          {
-            label: 'The Future of Media Commission',
-            value: '0f93b-the-future-of-media-commission',
-          },
-          {
-            label: 'The Housing Commission',
-            value: 'c78a8-the-housing-commission',
-          },
-          {
-            label: 'The State Laboratory',
-            value: '0cad3-state-laboratory',
-          },
-          {
-            label: 'Tipperary County Council',
-            value: 'tipperary-county-council',
-          },
-          {
-            label: 'Waterford City and County Council',
-            value: 'waterford-city-county-council',
-          },
-          {
-            label: 'Western Development Commission (WDC)',
-            value: '14371907-western-development-commission-wdc',
-          },
-          {
-            label: 'Westmeath County Council',
-            value: 'westmeath-county-council',
-          },
-          {
-            label: 'Wexford County Council',
-            value: 'wexford-county-council',
-          },
-          {
-            label: 'Wicklow County Council',
-            value: 'wicklow-county-council',
-          },
-        ],
-      },
-      {
-        label: 'Categories',
-        options: [
-          {
-            label: 'Agencies',
-            value: 'agencies',
-          },
-          {
-            label: 'Aquaculture licences',
-            value: 'aquaculture_licence',
-          },
-          {
-            label: 'Biographies',
-            value: 'biographies',
-          },
-          {
-            label: 'Campaigns',
-            value: 'campaigns',
-          },
-          {
-            label: 'Circulars',
-            value: 'circulars',
-          },
-          {
-            label: 'Collections',
-            value: 'collections',
-          },
-          {
-            label: 'Consultations',
-            value: 'consultations',
-          },
-          {
-            label: 'Coroner cases',
-            value: 'coroner_cases',
-          },
-          {
-            label: 'Departments',
-            value: 'departments',
-          },
-          {
-            label: 'Forms',
-            value: 'forms',
-          },
-          {
-            label: 'IGEES Publications',
-            value: 'igees_publication',
-          },
-          {
-            label: 'Local authorities',
-            value: 'local_authorities',
-          },
-          {
-            label: 'News',
-            value: 'news',
-          },
-          {
-            label: 'Organisation information',
-            value: 'organisation_information',
-          },
-          {
-            label: 'Policies',
-            value: 'policies',
-          },
-          {
-            label: 'Policy information',
-            value: 'policy_information',
-          },
-          {
-            label: 'Press releases',
-            value: 'press_releases',
-          },
-          {
-            label: 'Publications',
-            value: 'general_publications',
-          },
-          {
-            label: 'Reports',
-            value: 'reports',
-          },
-          {
-            label: 'Roles',
-            value: 'roles',
-          },
-          {
-            label: 'Services',
-            value: 'services',
-          },
-          {
-            label: 'Speeches',
-            value: 'speeches',
-          },
-        ],
-      },
-      {
-        label: 'Topic',
-        options: [
-          {
-            label: 'Brexit',
-            value: 'brexit',
-          },
-          {
-            label: 'Business and industry',
-            value: 'business-and-industry',
-          },
-          {
-            label: 'COVID-19',
-            value: 'covid-19',
-          },
-          {
-            label: 'Corporate information',
-            value: 'corporate-information',
-          },
-          {
-            label: 'Crime, justice and law',
-            value: 'crime-justice-and-law',
-          },
-          {
-            label: 'Defence and armed forces',
-            value: 'defence-and-armed-force',
-          },
-          {
-            label: 'Education, training and skills',
-            value: 'education-training-and-skills',
-          },
-          {
-            label: 'News',
-            value: 'news',
-          },
-          {
-            label: 'Organisation information',
-            value: 'organisation_information',
-          },
-          {
-            label: 'Policies',
-            value: 'policies',
-          },
-          {
-            label: 'Policy information',
-            value: 'policy_information',
-          },
-          {
-            label: 'Press releases',
-            value: 'press_releases',
-          },
-        ],
-      },
-    ],
-  },
-};
